+++ libsolidity/formal/CHC.cpp
@@ -1030,12 +1030,12 @@ Predicate const* CHC::createBlock(ASTNode const* _node, PredicateType _predType,
 	return block;
 }
 
-Predicate const* CHC::createSummaryBlock(FunctionDefinition const& _function, ContractDefinition const& _contract)
+Predicate const* CHC::createSummaryBlock(FunctionDefinition const& _function, ContractDefinition const& _contract, PredicateType _type)
 {
 	return createSymbolicBlock(
 		functionSort(_function, &_contract, state()),
 		"summary_" + uniquePrefix() + "_" + predicateName(&_function, &_contract),
-		PredicateType::FunctionSummary,
+		_type,
 		&_function
 	);
 }
@@ -1137,6 +1137,8 @@ smtutil::Expression CHC::predicate(Predicate const& _block)
 	case PredicateType::ConstructorSummary:
 		return constructor(_block, m_context);
 	case PredicateType::FunctionSummary:
+	case PredicateType::InternalCall:
+	case PredicateType::ExternalCall:
 		return smt::function(_block, m_currentContract, m_context);
 	case PredicateType::FunctionBlock:
 		solAssert(m_currentFunction, "");
@@ -1208,7 +1210,17 @@ smtutil::Expression CHC::predicate(FunctionCall const& _funCall)
 		args.push_back(currentValue(*var));
 	}
 
-	return (*m_summaries.at(calledContract).at(function))(args);
+	Predicate const& summary = *m_summaries.at(calledContract).at(function);
+	auto from = smt::function(summary, calledContract, m_context);
+	Predicate const& callPredicate = *createSummaryBlock(
+		*function,
+		*calledContract,
+		kind == FunctionType::Kind::Internal ? PredicateType::InternalCall : PredicateType::ExternalCall
+	);
+	auto to = smt::function(callPredicate, calledContract, m_context);
+	addRule(smtutil::Expression::implies(from, to), to.name);
+
+	return callPredicate(args);
 }
 
 void CHC::addRule(smtutil::Expression const& _rule, string const& _ruleName)
@@ -1464,6 +1476,9 @@ optional<string> CHC::generateCounterexample(CHCSolverInterface::CexGraph const&
 
 	auto callGraph = summaryCalls(_graph, *rootId);
 
+	auto nodePred = [&](auto _node) { return Predicate::predicate(_graph.nodes.at(_node).name); };
+	auto nodeArgs = [&](auto _node) { return _graph.nodes.at(_node).arguments; };
+
 	bool first = true;
 	for (auto summaryId: callGraph.at(*rootId))
 	{
@@ -1475,8 +1490,6 @@ optional<string> CHC::generateCounterexample(CHCSolverInterface::CexGraph const&
 		solAssert(stateVars.has_value(), "");
 		auto stateValues = summaryPredicate->summaryStateValues(summaryArgs);
 		solAssert(stateValues.size() == stateVars->size(), "");
-
-		string txCex = summaryPredicate->formatSummaryCall(summaryArgs);
 
 		if (first)
 		{
@@ -1504,7 +1517,23 @@ optional<string> CHC::generateCounterexample(CHCSolverInterface::CexGraph const&
 				path.emplace_back("State: " + modelMsg);
 		}
 
-		path.emplace_back(txCex);
+		string txCex = summaryPredicate->formatSummaryCall(summaryArgs);
+
+		list<string> calls;
+		auto dfs = [&](unsigned node, unsigned depth, auto&& _dfs) -> void {
+			auto pred = nodePred(node);
+			solAssert(pred && pred->isSummary(), "");
+			if (!pred->isConstructorSummary())
+				for (unsigned v: callGraph[node])
+					_dfs(v, depth + 1, _dfs);
+			calls.push_front(string(depth * 2, ' ') + pred->formatSummaryCall(nodeArgs(node)));
+			if (pred->isInternalCall())
+				calls.front() += " -- internal call";
+			else if (pred->isExternalCall())
+				calls.front() += " -- external call";
+		};
+		dfs(summaryId, 0, dfs);
+		path.emplace_back(boost::algorithm::join(calls, "\n"));
 	}
 
 	return localState + "\nTransaction trace:\n" + boost::algorithm::join(boost::adaptors::reverse(path), "\n");
@@ -1514,16 +1543,27 @@ map<unsigned, vector<unsigned>> CHC::summaryCalls(CHCSolverInterface::CexGraph c
 {
 	map<unsigned, vector<unsigned>> calls;
 
-	solidity::util::BreadthFirstSearch<pair<unsigned, unsigned>>{{{_root, _root}}}.run([&](auto info, auto&& _addChild) {
-		auto [node, root] = info;
-		if (Predicate::predicate(_graph.nodes.at(node).name)->isSummary())
+	auto compare = [&](unsigned _a, unsigned _b) {
+		return _graph.nodes.at(_a).name > _graph.nodes.at(_b).name;
+	};
+
+	queue<pair<unsigned, unsigned>> q;
+	q.push({_root, _root});
+	while (!q.empty())
+	{
+		auto [node, root] = q.front();
+		q.pop();
+
+		Predicate const* nodePred = Predicate::predicate(_graph.nodes.at(node).name);
+		if (nodePred->isSummary() && (_root == root || nodePred->isInternalCall() || nodePred->isExternalCall()))
 		{
 			calls[root].push_back(node);
 			root = node;
 		}
-		for (auto v: _graph.edges.at(node))
-			_addChild({v, root});
-	});
+		auto const& edges = _graph.edges.at(node);
+		for (unsigned v: set<unsigned, decltype(compare)>(begin(edges), end(edges), compare))
+			q.push({v, root});
+	}
 
 	return calls;
 }