+++ Changelog.md
@@ -9,6 +9,7 @@ Compiler Features:
  * SMTChecker: Support ABI functions as uninterpreted functions.
  * SMTChecker: Use checked arithmetic by default and support ``unchecked`` blocks.
  * SMTChecker: Show contract name in counterexample function call.
+ * SMTChecker: Output internal and trusted external function calls in a counterexample's transaction trace.
 
 Bugfixes:
  * Code Generator: Fix length check when decoding malformed error data in catch clause.